Load environment variables at the start of the tasks' process.
There are environment variables that control the behaviour of the R
interpreter, such as tuning the GC. These variables must be set at the
start at the process. Setting them from R code, as we do now, is too
late and has no effect.

This commit exports all non-secret environment variables into an
Renviron file in the task directory, which the batch script configures
using R_ENVIRON_USER. Once of the first things the R interpreter does is
load that file and set values it finds as environment variables.

The existing mechanism to load environment variables is kept. It is
necessary both to support secret variables, which we don't want to
expose in the Renviron, and to allow tasks to run without starting a new
process at all, as is commonly done in tests and in the example driver.

Retry tasks need a bit of care for this to work as well. For the most
part, the new task's folder only contains an indirection to the original
task's data and not a copy of it. When the task starts, it follows the
indirection to find the actual data. This strategy does not work for
the Renviron file, so a copy of that file is made.