Set R_GC_MEM_GROW=3 globally for all tasks.
We've found that setting this value provides significant speedup for
some of our workloads. The downside is a slight increase in memory
consumption, but our jobs are not usually memory bound anyway.

This can still be overriden, either by setting the
`hipercow.default_envvars` option, or by assigning a specific value on
individual tasks.