Fix the path to the status file in the batch file.
The batch file to run tasks checks for the presence of a
`status-success` file when exiting, and derives its exit code from that.
This is done to report useful status on the HPC web portal.

Unfortunately the path it was looking for was incorrect, therefore the
file never existed and all jobs are reported as failed. This fixes the
path by removing a spurious `log\` component.