Print environment contents when loading it.
hipercow silently attaches packages and sources files from the
configured environment. If something goes wrong in this process (eg. the
file does not exist), it can be difficult for users to understand what
is happening. None of their own code makes any mention of the files or
packages in question, and they may have forgotten that they ever created
the environment.

This prints a message when loading the environment, with the list of
files and packages that will be loaded. This should provide enough hints
for the user to make the connection between a hypothetical error and the
environment.