Allow rrq tasks to be started from a hipercow task.
The `use_rrq` option was broken: it is supposed to initialize an rrq
controller in the task's process, but in order to do so it needs to find
the queue ID. There was code to read a queue ID from the environment,
but nowhere relevant was that environment variable set. In theory it
could read it from disk, but the `HIPERCOW_NO_DRIVERS` option did not
allow that.

The tests for that functionality were passing because they were
manually setting `HIPERCOW_RRQ_QUEUE_ID`.

This now passes the queue ID as part of the `parallel` argument to the
task, fixing the issue. Added a more comprehensive test, which actually
starts the rrq tasks from a hipercow task.