Configure the rrq offload directory each time a controller is created.
We previously setup the rrq offload directory once when creating the
queue, storing the absolute path to it in the redis database. Using a
single global value breaks down when that directory is mounted on
multiple machines but at differing paths.

The most spectacular failure case happens when configuring the queue
from a linux machine and using it to run on a Windows cluster. When this
happens, the offload path looks something like
`/home/<user>/path/to/project/hipercow/rrq/offload`, which when used
from a Windows machine obviously does not work.

This is solved by configuring each instance of an `rrq_controller` with
its own path to the offload directory, relative to the hipercow root.
This root is already automatically adjusted based on where we are
running.